flask/requirements.txt
Flask
gunicorn
Flask-Uploads
Pillow
pathlib
Werkzeug
jsonify
requests